98% done

+++ src/routes/Contacts.js
@@ -66,17 +66,36 @@ router.post("/add", validateToken, async (req, res) => {
 router.post("/delete", validateToken, async (req, res) => {
   try {
     const datas = req.body;
-
-    const data = await contacts.find({ userid: req.user });
-    if (data) {
-      datas.map(async (ids) => {
-        await contacts.deleteOne({ _id: ids.id });
-      });
-      res.json({ message: "success", data });
+    if (datas.length > 0) {
+      const data = await contacts.find({ userid: req.user });
+      if (data) {
+        datas.map(async (ids) => {
+          await contacts.deleteOne({ _id: ids.id });
+        });
+        res.json({ message: "success", data });
+      }
+    } else {
+      res.json("not deleted");
     }
   } catch (e) {
     res.status(400).json({ message: err.message });
   }
 });
 
+router.post("/sdelete", validateToken, async (req, res) => {
+  try {
+    const datas = req.body;
+    if (datas[0].id != "" && datas.length > 0) {
+      const data = await contacts.find({ userid: req.user });
+      if (data) {
+        datas.map(async (ids) => {
+          await contacts.deleteOne({ _id: ids.id });
+        });
+        res.json({ message: "success", data });
+      }
+    }
+  } catch (e) {
+    res.status(400).json({ message: err.message });
+  }
+});
 module.exports = router;
